pyproject.toml
[project]
name = "god_eye"
version = "1.0.0"
description = "GOD_EYE — Open Source OSINT Intelligence Platform for digital footprint analysis and privacy auditing"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "GOD_EYE Contributors" }]
keywords = ["osint", "privacy", "security", "intelligence", "digital-footprint"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    # CLI & UI
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",

    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",

    # Async HTTP
    "aiohttp>=3.9.3",
    "aiofiles>=23.2.1",

    # Browser Automation
    "playwright>=1.41.0",

    # Databases
    "neo4j>=5.17.0",
    "redis[hiredis]>=5.0.1",
    "aiosqlite>=0.20.0",

    # Data Validation & Settings
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # DNS & Network
    "dnspython>=2.6.0",
    "python-whois>=0.9.4",

    # Phone
    "phonenumbers>=8.13.31",

    # Image Processing
    "Pillow>=10.2.0",
    "exifread>=3.0.0",
    "imagehash>=4.3.1",

    # AI & NLP
    "anthropic>=0.40.0",
    "openai>=1.12.0",
    "spacy>=3.7.4",

    # Face Recognition (optional install — heavy dependency)
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    "numpy>=1.26.4",

    # Export
    "jinja2>=3.1.3",
    "markdown>=3.5.2",
    "weasyprint>=62.3",

    # Scraping & Parsing
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
    "tenacity>=8.2.3",
    "python-dateutil>=2.9.0",
    "click>=8.1.7",
    "tabulate>=0.9.0",
    "geopy>=2.4.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
]
dev = [
    "god_eye[test]",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.12",
    "types-beautifulsoup4>=4.12.0",
    "types-aiofiles>=23.2.0",
    "ipython>=8.20.0",
    "pre-commit>=3.6.0",
    "bandit[toml]>=1.7.7",
]
prod = []
nlp = [
    "spacy>=3.7.0",
]
ml = [
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.17.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
]

[project.scripts]
god_eye = "app.cli:app"
god-eye = "app.cli:app"

[project.urls]
Homepage = "https://github.com/yourorg/god_eye"
Documentation = "https://github.com/yourorg/god_eye/blob/main/README.md"
Repository = "https://github.com/yourorg/god_eye"
Issues = "https://github.com/yourorg/god_eye/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = ["E501", "B008", "B904"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=app --cov-report=term-missing"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
skip_empty = true
